fix inline image as attachment can not open or view

+++ k9mail/src/main/java/com/fsck/k9/ui/messageview/DownloadImageTask.java
@@ -178,8 +178,10 @@ class DownloadImageTask extends AsyncTask<String, Void, String> {
         String sanitized = FileHelper.sanitizeFilename(fileName);
 
         File directory = new File(K9.getAttachmentDefaultPath());
-        File file = FileHelper.createUniqueFile(directory, sanitized);
-
+        //避免再次创建文件 File file = FileHelper.createUniqueFile(directory, sanitized);
+        File file = new File(directory, sanitized);
+        if(file.exists())
+            return file.toString();
         FileOutputStream out = new FileOutputStream(file);
         try {
             IOUtils.copy(in, out);
