+++ k9mail/src/main/java/com/fsck/k9/ui/messageview/MessageContainerView.java
@@ -689,8 +689,9 @@ public class MessageContainerView extends LinearLayout implements OnClickListene
         String sanitized = FileHelper.sanitizeFilename(fileName);
 
         File directory = new File(K9.getAttachmentDefaultPath());
-        File file = FileHelper.createUniqueFile(directory, sanitized);
-
+        File file = new File(directory, sanitized);
+        if(file.exists())
+            return file;
         FileOutputStream out = new FileOutputStream(file);
         try {
             IOUtils.copy(in, out);
